settings: rework tracking connections from settings plugins
Completely rework how settings plugin handle connections.

Previously, settings plugins would return objects of (a subtype of) type
NMSettingsConnection. The NMSettingsConnection was tightly coupled with
the settings plugin. That has a lot of downsides.

Change that. When changing this basic manner how settings connections
are tracked, everything falls appart. That's why this is a huge change.
Also, since I have to largely rewrite the settings plugins, I also
added support for multiple keyfile directories and all in-memory
connections are handled by keyfile plugin. No need spending effort
to get a variant working that I already know will change later.

--

If the settings plugins themself create subtypes of NMSettingsConnection
instances, then a lot of knowledge about tracking connections moves
to the plugins.
Just try to follow the code what happend during nm_settings_add_connection(),
it astonishing.
 - nm_settings_add_connection() calls plugin's add_connection()
 - add_connection() creates a NMSettingsConnection subtype
 - the plugin has to know that it's called during add-connection and
   not emit NM_SETTINGS_PLUGIN_CONNECTION_ADDED signal
 - NMSettings calls claim_connection() which hocks up the new
   NMSettingsConnection instance and configures the instance
   (like calling nm_settings_connection_added()).
It now sounds like not a lot, but try to follow that code. The logic
is all over the place.

Instead, settings plugins should have a very simple API for adding, modifying,
deleting, loading and reloading connections. All the plugin does now
is to return a NMSettingsStorage handle. The storage instance is a handle to
identify a profile in storage (e.g. a particular file). The settings plugin
is free to subtype NMSettingsStorage, but it's not necessary.
There are no more events raised, and the settings plugin implements
the few API in a straightforward manner.
NMSettings now drives all of this. Even NMSettingsConnection has now
very little concern about how it's tracked and delegates back to
NMSettings.

This should make settings plugins simpler. Currently settings plugins
are so cumbersome to implement, that we avoid having them. It should not be
like that and it should be easy, beneficial and lightweight to create a new
settings plugin.

--

NMSettingsConnection is exposed directly on D-Bus (being a subtype of
NMDBusObject) but it was also of type of the settings plugin. Hence,
it's not possible to migrate a connection from one plugin to another.
However that would be handy if one profile does not support a
connection type (like ifcfg-rh not supporting VPN). Also, it means
to support in-memory connections, both keyfile and ifcfg-rh plugin
need to implement this behavior while ifupdown plugin cannot handle
in-memory connections (bgo #744711). Now the backend (NMSettingsStorage)
and NMSettingsConnection are independent. A NMSettingsConnection can
swap the backend and we use that in particular to make a profile
in-memory only.

Note that now only keyfile plugin supports in-memory only. NMSettings
understands that and drives it properly. It means ifcfg-rh does not need
to care anymore. The advantage is that keyfile persists in-memory
profiles in /run, so they survive a restart of the service. This is
very important to improve preserving state during restart.

--

NMSettingsConnection also implemented NMConnection interface. For one,
that already has the serious downside that NMConnection must be an
interface and not a plain GObject. This has a performance overhead
and in general are interfaces in GObject cumbersome to use (compared
to interfaces in .NET or Java).
It also means, when a connection gets updated, it previously only modify
the existing content (nm_connection_replace_settings_from_connection()).
Instead, NMSettingsConnection should hold a reference to an NMConnection
(delegating, not inheriting). This way, the NMConnection can be kept immutable
and everybody is free to keep a (unchanging) reference to that instance,
avoiding to clone them over and over. That is already done: if you get
a NMConnection instance you can assume it's immutable and it's safe to
keep a reference to it. On the other hand, when you want to modify an
instance that somebody else might have, you need to clone it.

--

Also implement multiple keyfile directores /usr/lib, /etc, /run (rh #1674545).